requirements.txt
openai==1.54.0
python-dotenv==1.0.1
invoke==2.2.0
pydantic==2.9.1
azure-identity==1.19.0
pytest==8.3.3
coverage==7.6.1
pytest-cov==5.0.0
fastapi==0.115.5
uvicorn==0.32.0
grpcio-tools==1.66.2
grpcio-reflection==1.67.0
starlette_gzip_request==0.1.0
pdoc==15.0.0
llmlingua==0.2.2
pydocstyle==6.3.0
flake8==7.1.1
flake8-docstrings==1.7.0
docformatter==1.7.5
black===24.10.0